requirements.txt
discord
pymongo[srv]
python-dotenv
pyyaml
